flake.nix: substitute nvidia-smi path in ollama sources
Previous attempt to include linuxPackages_latest.nvidia_x11 in the
ollama buildInputs was wrong. The nvidia-smi is a runtime dependency of
ollama, which expected it to be found in the PATH. Adding nvidia_x11 to the
buildInputs did not satisfy that expectation.
Alternative solution I chose was to patch the ollama sources to
reference the nvidia-smi nix store path explicitly. Ollama binary that
references the nvidia-smi nix store path is then recognized by Nix as
depending on the linuxPackages_latest.nvidia_x11 package, and the latter
is included in the ollama closure.

This inclusion can be verfied by:
> nix-store -q --references /nix/store/6bv2bk4xmw71s9pnlbj5xjnzbc9n76s3-ollama-0.1.7
/nix/store/5gq0ip7xal3r0789qa4a0wr51k8igxhq-mailcap-2.1.53
/nix/store/cf0gxd6nyr6shkkn0ca82bap1x8r0f9x-tzdata-2023c
/nix/store/gm4cza2xl5frhkm86j3qclmsqh7ivqxb-iana-etc-20230316
/nix/store/gqghjch4p1s69sv4mcjksb2kb65rwqjy-glibc-2.38-23
/nix/store/yi3a0mc6kam5zrxdfdgr71zjv3akywsr-nvidia-x11-545.29.02-6.1.61
/nix/store/z24x9shlz5minif7vcifxipgazjp3pl5-llama-cpp-1469