remove the dependency on linuxPackages.nvidia_x11.bin
nvidia-smi depends on a specific version of the installed nvidia driver.
Specifying that it depends on linuxPackages.nvidia_x11.bin may cause
a mismatch between the nvidia driver and nvidia-smi.

I don't have a good solution for this problem other than explicitly
warning the user that they should have nvidia-smi available in the PATH.